Create kitoboot.py
Kitoboot is the boot image processing tool for the beta version of Kito-android and kernbox2. In the stable version of Kito-android, all functions have been implemented using Rust, so this boot image processing tool is no longer maintained or optimized by us. Therefore, it may have some potential issues! However, it seems to work normally when running the Unpack/Repack functions in a regular environment. Kitoboot is implemented by referring to magiskboot and mkbootimg, so you can consider Kitoboot as a Python-written magiskboot.